+++ miden-crypto/src/hash/algebraic_sponge/poseidon2/mod.rs
@@ -35,12 +35,11 @@ mod test;
 /// and it can be serialized into 32 bytes (256 bits).
 ///
 /// ## Hash output consistency
-/// Functions [hash_elements()](Poseidon2::hash_elements), [merge()](Poseidon2::merge), and
-/// [merge_with_int()](Poseidon2::merge_with_int) are internally consistent. That is, computing
-/// a hash for the same set of elements using these functions will always produce the same
-/// result. For example, merging two digests using [merge()](Poseidon2::merge) will produce the
-/// same result as hashing 8 elements which make up these digests using
-/// [hash_elements()](Poseidon2::hash_elements) function.
+/// Functions [hash_elements()](Poseidon2::hash_elements), and [merge()](Poseidon2::merge), are
+/// internally consistent. That is, computing a hash for the same set of elements using these
+/// functions will always produce the same result. For example, merging two digests using
+/// [merge()](Poseidon2::merge) will produce the same result as hashing 8 elements which make up
+/// these digests using [hash_elements()](Poseidon2::hash_elements) function.
 ///
 /// However, [hash()](Poseidon2::hash) function is not consistent with functions mentioned above.
 /// For example, if we take two field elements, serialize them to bytes and hash them using
@@ -162,12 +161,6 @@ impl Poseidon2 {
     #[inline(always)]
     pub fn merge_many(values: &[Word]) -> Word {
         <Self as AlgebraicSponge>::merge_many(values)
-    }
-
-    /// Returns a hash of a digest and a u64 value.
-    #[inline(always)]
-    pub fn merge_with_int(seed: Word, value: u64) -> Word {
-        <Self as AlgebraicSponge>::merge_with_int(seed, value)
     }
 
     /// Returns a hash of two digests and a domain identifier.