+++ miden-crypto/src/hash/algebraic_sponge/rescue/rpo/mod.rs
@@ -29,11 +29,10 @@ mod tests;
 /// and it can be serialized into 32 bytes (256 bits).
 ///
 /// ## Hash output consistency
-/// Functions [hash_elements()](Rpo256::hash_elements), [merge()](Rpo256::merge), and
-/// [merge_with_int()](Rpo256::merge_with_int) are internally consistent. That is, computing
-/// a hash for the same set of elements using these functions will always produce the same
-/// result. For example, merging two digests using [merge()](Rpo256::merge) will produce the
-/// same result as hashing 8 elements which make up these digests using
+/// Functions [hash_elements()](Rpo256::hash_elements), and [merge()](Rpo256::merge), are internally
+/// consistent. That is, computing a hash for the same set of elements using these functions will
+/// always produce the same result. For example, merging two digests using [merge()](Rpo256::merge)
+/// will produce the same result as hashing 8 elements which make up these digests using
 /// [hash_elements()](Rpo256::hash_elements) function.
 ///
 /// However, [hash()](Rpo256::hash) function is not consistent with functions mentioned above.
@@ -146,12 +145,6 @@ impl Rpo256 {
     #[inline(always)]
     pub fn merge_many(values: &[Word]) -> Word {
         <Self as AlgebraicSponge>::merge_many(values)
-    }
-
-    /// Returns a hash of a digest and a u64 value.
-    #[inline(always)]
-    pub fn merge_with_int(seed: Word, value: u64) -> Word {
-        <Self as AlgebraicSponge>::merge_with_int(seed, value)
     }
 
     /// Returns a hash of two digests and a domain identifier.